# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AuditEnvelope:
    properties:
      logs:
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.PageMeta'
    type: object
  dto.AuthResponse:
    properties:
      message:
        example: Logged in
        type: string
      refresh_token:
        example: 9f2c4a1e6b4d4f8a9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f5a6b7c8d9e0f1a2
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3OCJ9.dozjgNryP4J3jVmNHl0w5N_XgL0n3I9PlF
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  dto.ErrorEnvelope:
    properties:
      error:
        type: string
    type: object
  dto.ForgotPasswordRequest:
    properties:
      email:
        example: kaige@example.com
        type: string
    required:
    - email
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        example: kaige@example.com
        type: string
      password:
        example: supersecret123
        maxLength: 72
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
...
  dto.MessageEnvelope:
    properties:
      message:
        type: string
    type: object
...
  dto.PageMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.RefreshRequest:
    properties:
      refresh_token:
        example: 9f2c4a1e6b4d4f8a9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f5a6b7c8d9e0f1a2
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterRequest:
    properties:
      age:
        example: 30
        maximum: 100
        minimum: 18
        type: integer
      email:
        example: kaige@example.com
        type: string
      name:
        example: Kaige Saif
        maxLength: 50
        minLength: 3
        type: string
      password:
        example: supersecret123
        maxLength: 72
        minLength: 8
        type: string
    required:
    - age
    - email
    - name
    - password
    type: object
  dto.ResendVerificationRequest:
    properties:
      email:
        example: kaige@example.com
        type: string
    required:
    - email
    type: object
  dto.ResetPasswordRequest:
    properties:
      new_password:
        example: newsecret123
        maxLength: 72
        minLength: 8
        type: string
      token:
        description: Token may alternatively be supplied as ?token= query param.
        example: 9f2c4a1e6b4d4f8a9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f5a6b7c8d9e0f1a2b
        type: string
    required:
    - new_password
    type: object
...
  dto.VerifyOTPRequest:
    properties:
      code:
        example: "482914"
        type: string
      email:
        example: kaige@example.com
        type: string
    required:
    - code
    - email
    type: object
  models.AuditLog:
    properties:
      action:
        example: contacts.create
        type: string
      actor_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      ip:
        example: 203.0.113.7
        type: string
      metadata:
        additionalProperties: {}
        type: object
      org_id:
        type: string
      request_id:
        type: string
      resource_id:
        type: string
      resource_type:
        example: contacts
        type: string
    type: object
...
  models.Profile:
    properties:
      address:
        example: 123 Main St, Lagos
        type: string
      avatar_url:
        example: https://example.com/avatars/kaige.png
        type: string
      bio:
        example: Gopher since 2021. Coffee first.
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        type: string
      phone:
        example: "+15551234567"
        type: string
      updated_at:
        type: string
      user_id:
        example: 458622d8-daba-4252-8ce1-846277353139
        type: string
    type: object
  models.User:
    properties:
      age:
        example: 30
        type: integer
      created_at:
        type: string
      deleted_at:
        description: |-
          Soft delete: Delete() sets this instead of removing the row.
          Queries filter it out automatically; uniqueness is partial
          (WHERE deleted_at IS NULL) so values stay reusable.
        type: string
      email:
        example: kaige@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: 458622d8-daba-4252-8ce1-846277353139
        type: string
      name:
        example: Kaige Saif
        type: string
      profile:
        $ref: '#/definitions/models.Profile'
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Grip API with JWT auth and email verification.
  title: Grip API
  version: "1.0"
paths:
  /audit-logs/:
    get:
      parameters:
      - description: Filter by actor user ID
        in: query
        name: actor_id
        type: string
      - description: Filter by action
        example: contacts.create
        in: query
        name: action
        type: string
      - description: Filter by resource
        example: contacts
        in: query
        name: resource_type
        type: string
      - description: RFC3339 start
        example: "2026-01-01T00:00:00Z"
        in: query
        name: from
        type: string
      - description: RFC3339 end
        example: "2026-12-31T23:59:59Z"
        in: query
        name: to
        type: string
      - default: 1
        description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        example: 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuditEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List organization audit logs (admin+)
      tags:
      - audit
  /audit-logs/{id}:
    get:
      parameters:
      - description: Audit log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuditLog'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Get one organization audit log (admin+)
      tags:
      - audit
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      summary: Request a password-reset email
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      summary: Log in with email and password
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      summary: Revoke one refresh token (logout)
      tags:
      - auth
  /auth/logout-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Revoke all refresh tokens (logout everywhere)
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Get the authenticated user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      summary: Rotate a refresh token into a new pair
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      summary: Register a new user
      tags:
      - auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      summary: Resend verification email
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      summary: Reset password with a reset token
      tags:
      - auth
  /auth/verify:
    get:
      parameters:
      - description: Verification token from email
        example: 1ef700f6a95416bb2a1d90a6ee3e8d62e78f9b3e084e405ab29fd722a57c96e8
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      summary: Verify email address
      tags:
      - auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email and 6-digit code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      summary: Verify email with OTP code
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT. Example: "Bearer
      eyJhbGciOiJIUzI1NiJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"